pyproject.toml
[project]
name = "python-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all,standard]>=0.115.12",
    "langchain-groq>=0.3.2",
    "mcp-use>=1.2.7",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
    "starlette>=0.46.2",
]
